requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
schedule==1.2.0
websocket-client==1.6.4
pandas==2.1.3
numpy==1.25.2
pytest==7.4.3
black==23.10.1
flake8==6.1.0
mypy==1.7.0
cryptography==41.0.7